chore(suite): remove build tags
The integration suite was guarded behind the "suite" build tag, which
prevented Go tooling from working with the integration tests by default.
This made the upcoming refactoring to support API tokens more difficult,
and has historically discouraged the addition of more integration tests.

As modern versions of Go support `testing.Cleanup` we can move the
envtest setup into a helper function that runs, and automatically cleans
up, in every test, instead of wrapping the main function. The tests are
guarded by checking for the presence of the KUBEBUILDER_ASSETS
environment variable, which is set by `setup-env` in the CI environment.