fix(certificaterequest): ignore empty issuer group
A CertificateRequest's reference to an Issuer can have an empty Group,
which we were considering as equivalent to matching, then later
rejecting the CertificateRequest when it was for an unknown issuer.
While setting the group is optional in the schema, our documentation has
always included the group in the issuer reference. The rejection was
racing with other controllers, including the built-in controllers, and
preventing them from issuing certificates.

This changeset updates CertificateRequestController to ignore
CertificateRequests where the issuer group is unset, preventing the
race with controllers for other issuers. We'll continue to reject
CertificateRequests for unknown kinds if the group matches our group.

Bug: #149
Fixes: d8c1b77 ("feat: add ClusterOriginIssuer")